Staretd work on the project two class
None of the methods have actually been touched yet, but all of their javadoc is fully written.

The HW2Tests class has been created and imports the entire assert branch of junit as a static method, so that tests can be done easily. The tests are already named based on what method they are testing from the parent class.

The HW2Main class has been created to allow the user to test the methods of the parent through the cmd line. While this could've easily been placed in the parent class, I chose to place it in its own file that extends the parent to better fit the needs of the project and wants of the professor. The techniques I used in the main class have not been formally taught, and we have been advised to do no testing in the main method, so this class should not be included in ANY project submissions.

Good luck on the project, I hope you can figure out some of the methods before youre crunched for time with upcoming exams.

2/18/24 @ 12:49pm