Finished 4/6 methods of Project 2
Of the required 6 methods, I completed all but the 'flipEachK' and 'replaceText' Methods. All testing still needs to be performed, and they should be performed before attempting to code the last method and the bonus method.

All of the methods appear to work flawlessly, though more rigorous testing that is not done in the command line must be done to ensure perfection and a good grade.

2/18/24 @ 6:56pm