Redid HW2 Method 6 and Lecture Code
HW2 Method 6 is not working as intended. It appears that it is ignoring any possible replacement strings that contain more than one replacement. Edge cases need to be tested.

IMPORTANT: the code must be changed significantly to allow for nested parentheses. i.e the code must only replace a substring if there is a fully closed parentheses statement.

((hello))) should replace ((hello)) and leave the ).

(hello worlds) with (spy(me)) should replace the base string with spy(me)

This will be logically hard to figure out. Go slow and test as you code.

2/21/24 @ 3:07pm