Worked on test cases in the first three methods
Work on organizing the tests to follow the 0, 1, many and first, middle, last framework outlined in the project instructions.

For the 6th method, I placed Aprils code in the obsidian file for the project,and placed an algorithm in the instructions that I wrote myself. Im still slightly confused but things are atarting to make more sense.

Finish testing for the first 5 methods before you start on the 6th method please. You will thank me later.

2/23/24 @ 3:05pm