Added second helper method and did Lecture code
Second helper method exists only to loop for parentheses in an inputted string to reduce any redundancies. This is used twice in the project, once in matching parentheses method and once at the start of replaceText. It is necessary to  have a method for this to save space doing a mundane task.

Lecture code also updated, some notes in obsidian.

2/26/24 @3:07pm