Worked on Method 6
Progress has been made. Currently I have only done searching on the base String.

If an open parentheses is found, then the program starts to look for a matching closed parentheses. If another open parentheses is found then it is indicated using a boolean variable. If the boolean variable is true than the program starts to search through the baseString to look for nested parentheses. If nested parentheses are found, then the start and end of the substring containing them is saved. If nested parentheses are not found, then the program remembers the start and of the substring as the instance of the mismatch and the temporary variable that stores the index of the closing parenthesis.

Next steps...

indicate whether or not a replacement is going to happen using a boolean variable. If this is true, then the large while loop should break into a branch where It searches through the replacements using a similar method as the previous paragraph. If the replacement variable is false, then the code should go into the code that I described above.

What I am looking for...

The large while loop should ONLY APPEND ONE REPLACEMENT CHARACTER AT A TIME. Some of the nested loops also append characters, and that is ok because they are compensating for mismatches.

If replacements will happen, then a replacement must be found. If a replacement cannot be found, then the indicated substring in the base is ignored from the StringBuilder and the index of the base moves on to the next character outside of the substring.

Good luck implimenting thhese things, I wrote a lot to try and assist you in the creation of the rest of the method because I wont look at it again for a few hours.

2/26/24 @ 4:53pm