Summary of HW2 and Next Steps...
All of the tests have passed and the code is functional, despite being very long. It might be worth clarifying if helper methods need tests or not, but besides that all of the actual code is complete.

As for the obsidian file, Reread a lot of the tests for replaceTexts, and make sure they are up to date. Mostly look at the use of spaces in the expected part of the assertions. They have been updated in the actual code blocks, but maybe some of the wording is outdated.

What do I mean by spaces?
I did not anticipate the spaces around substrings being included when a substring is dropped, but I shouldv'e expected this to happen. Nowhere does it say this behavior can not be exhibited, and it makes sense that it happens because the ignoring of a substring from base when replacements runs out is simply ignoring those characters, not interfereing with the spaces around it. This can be said for every other replacement done.

Helper methods...
I wrote, to my knowledge, 5 the entire time. I am unsure if they need to be fully tested, but it doesnt indicate that they do. Please make sure they are private before submitting. One of the helper methods is very long but it needs to be in order for the code to work properly.

Overall, the logic of the 6th method definetly was harder than I expected, but we got there in the end. It may have been frustrating to work through and debug, but I got there, adn I learned a lot because of it.

Good luck on the midterm next monday! In the mean time copy over the practice questions and review the weekly quizzes (maybe make a quizlet or find one online too).

Probably my longest commit to date, but hopefully theres something useful.

TLDR; Make sure the obsidian is up to date, ask about helper method tests (if needed), and submit project by friday while also studying for the midterm on the 4th of March.

2/28/24 @ 1:31am