Removed main method from HW2 class
Feedback recieved on the first project. Was told to not include mian method in submission which is fine. Main method in HW2 is in its own class. Review StudentAccount's charge() method. Was told that it does not set balance to zero after charging the difference.

Congrats on the first project and the praise you recieved for doing credit() correctly in the StudentAccount class.

2/29/24 @6:31pm