Altered helper method and calls
nestedParentheses method changed so that the inital sum is just equal to initial amount entered in the part of the call. Because of this, the calls in the body of the class were changed so that they all have an initial value of 1 instead of 0.

The old method had 'sum = 1 + initial' but that was not a good way to write the mehtod. This is because the sum could be at most 1 assuming initial is a positive or 0 integer. Chnaging it to the way described in the section above allows all possible values of 'inital' to be entered and be accurately reflected in the sum. The only modifications done to the code are the changing of the above statement and the calls in the replaceText method.

All calls of the nestedParentheses method were changed to one becuase they are called only when there is a parentheses found, indicating an intial value of 1. This is reflected in the rest of the code.

Congrasts on making it through the end of this project. The changes have been submitted along with the testing document and the JUnit class.

3/1/24 @ 1:40pm