Finished everything up to the more complex funcitons
Class hierarchy is looking very good, unary function has been created for complex functions listed above.

I plan on finishing the testing for the completed classes tonight, and I will finish the final methods in the next few days.

The methods are not too complex, and enums are used in both the unary and binary classes. There are a few things I want to try out in the unary abstract class, as Im not sure if its default behaviors are necessary. Extensive testing on the derivative and value methods should be done.

Make sure testing is consice, however, and indicate properly on the testing document without making it 20 pages long this time hehe.

Im curious about the polynomial class and recursively calling overriden methods such as toString, derivative etc.

Good luck on the rest of the project.

Haunted Mansion is a silly little movie.

Testing update later tonight, most-likely!