done and submitted
Final project is over! congrats on all your hard work this semester. The project is due Sunday 4/21, so there si time to refine anything in the main class or in the testing class/document. Also remeber to post any additional changes to the canvas comments and in the first part fo the testing document.

Something extra you might want to implement is a lose condition.